Use the archive download URL returned by Github's API
This fixes #142. We use the returned browser_download_url instead of
generating it ourselves. This fixes the issues with sanitized URLs and
filenames witout having to maintain sanitization code in `dune-release`.

The URL is stored in a `_build/<distrib_archive_base_filename>-url` file
when running `dune-release publish` during the distribution pubishing.

Later on, during `dune-release opam pkg`, if that file exists we use the
URL it contains otherwise we fallback to the old way. That makes it
easier to keep working with custom delegates for now, although I'm not
sure we really want to support them for long.