Fix documentation publication
When I replaced the hand-rolled `cp` code in documentation publication
I did not account for different `cp` behaviours depending on the
existence of the destination directory. That means that when one wants
to publish to the root of its gh-pages, the `_html` dir was copied as is
in there instead of its content because the root of the gh-pages
temporary project exists. This bug wasn't triggered during my tests
because I published the doc in a `doc` sub dir.