Make description optional for opam submit
Fixes #164

This makes the description field optional when submitting a PR to
opam-repository.

Given that opam considers it okay as long as there's a synopsis I think
it make sense to allow a missing description field. We might even
consider make that just a warning in the lint phase as well but that's a
separate matter I'd say. The important is that it doesn't prevent one
from releasing!