Distrib: Push the tag only if necessary
Check that the tag is already present on the remote and point to the
same rev as the local tag.

Pushing the tag may fail because the guessed URI may be wrong or git may
not have access to keys necessary for authentication.

This allows users to push the tag themselves.