Command `tag`: improve error and log messages
Before, when the `dune-release` command `tag` resulted in a git
command error (either because of trying to create a tag that already
existed or because of trying to delete one that didn't exist), dune-release
would just raise an error with git's error code and the command that had
failed.

With this commit, it checks why it failed and depending on that gives
an informative log message or an informative error messages. E.g.,
if the tag (default: version number) already exists and points to the
provided commit (default: HEAD), it informs the user that it aready exists.

Furthermore, if the user wants to delete a tag, but the tag provided
points to a different commit than the one provided, now dune-release
gives a warning when asking for confirmation.