Simplify distrib tarball preparation
This used to rely on Distrib.t which was in fact never instantiated
to anything than the default value. This simplifies this module to a
static list of paths to exclude for now.

Eventually we'll move some `Pkg` functions there but for now I mostly
wanted to get rid of all the indirections.