Fix a bug excluding all root dotfiles from the generated distrib tarball
`Bos.OS.Dir.fold_contents` behaviour though following its spec is very
counter intuitive since specifying `~dotfiles:true` will only include
dotfiles and folders in subdirectories.

Luckily this hasn't broken anything yet but it is obviously a bad
behaviour that needed fixing. The fact that we maintain an explicit
list of files and folders to exclude from the archive, some of which are
root dotfiles and folders confirms that this behaviour was not
intentional.