Suggest a solution for users without ssh set up
Users who haven't set up ssh will get a Permission denied error from
github when using `dune-release publish`. This commit adds to the README
how the user can configure git in order to solve that problem. It also
adds a hint to the Permission denied error message pointing to that
new part of the README.