Fix the process of inferring user name and repo
This commit fixes the following two minor bugs in the workflow of
inferring the user name and repo from the opam file fields `homepage` or
`dev-repo`.

- It adds a test case for the case the user puts a local path or an empty
string into the opam file fields. Before the commit, if the user did so,
the first two directories of the corresponding absolute path got inferred
as user and repo.

- Before this commit, the uri's in `homemapge` and `dev-repo` fields were
transformed into the corresponding uri's to the distribution by appending
"releases/download/{version}/{project name}-{version}.tbz" to them. From
these distribution uri's, user name and repo were inferred. So, if the
uri in the opam file fields was "https://foo.com/", then the process of
inferring user name and repo would not return an error, but infer
"releases" and "downloads" as user and repo, respectively. With this commit
user and repo get inferred directly from the uri's without first transforming
them into the distribution uri's.