Improve build/test logging
Writing "package" in singular misleads into thinking that only the main
package gets checked for build and tests, whereas in fact either all
packages or all packages passed in via `-p` are checked.

Particularly for the `check` command it's important to be precise here,
since in that command some checks are done just for the main package and
others for all packages.