Improve GH token management
The CLI option and env var used to take path which made them unusable
in a CI context. They now take the token value directly.

This also improves how the token is passed around, moving the config
reading bits as close to the CLI entry points as possible so the rest
of the code doesn't have to deal with this and simply require a
mandatory token argument, with the value and not the path.

This finally fixes a bug where the content of token files wasn't always
sanitized.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>