Check for the tag to be a tag, not just a valid ref
The code just checked the ref to exist which checked whether the ref
existed. When passing in an (existing) commit hash this would return
true, because a commit hash is a valid ref, but it is not a tag. This
changes the code so it actually tests that the passed reference is
indeed a tag.