Remove the path test cases
The logic is quite odd and the expected results are really misleading.
Some of them are impossible to test as-is because the tags would need to
be escaped but without running in a VCS repo it is impossible to pick
the right escaping code path so it is all just quite bad.