Make the order of packages in the PR description deterministic
Given the order is determined by the order the file system returns the
packages, it can be nondeterministic, different people might thus end up
with different orders and different file systems can cause different
orders.

This is not a big issue but resorting the packages alphabetically for
the title might lead to nicer/more consistent PRs where the title is
easier to scan.