Conflict with old version of result
Some packages pull in `result` as transitive dependency, resulting in
`Result` being not the Stdlib one but one that does not feature
`Result.is_ok` if the `result` package is too old.

Given we do not have a direct dependency this is added as a conflict, so
if our transitive dependencies drop `result` this problem should go
away without introducing spurious dependendencies.